tweaks + bugfixing

+++ source/nucledian/engine/graphics/shaders/billboard.frag
@@ -7,11 +7,11 @@ in vec3 position;
 in vec3 normal;
 in vec2 uv;
 
-layout(location = 0) out vec4  g_position;
-layout(location = 1) out vec4  g_normal;
-layout(location = 2) out vec3  g_stitched_normal;
-layout(location = 3) out vec4  g_albedo;
-layout(location = 4) out float g_sector;
+layout(location = 0) out vec4 g_position;
+layout(location = 1) out vec4 g_normal;
+layout(location = 2) out vec3 g_stitched_normal;
+layout(location = 3) out vec4 g_albedo;
+layout(location = 4) out uint g_sector;
 
 layout(binding = 0) uniform sampler2D sampler;
